much functionality so wow
got adding and deleting
got saving to spit out all the list items to a separate tab in a list
grid style... needs finagling. i also need to set the index to
display... or i need to create a window at the index +1 to the last
index that was created... gonna need to store the widget switcher
current last index somwhere for this purpose. the back button that i
create on this window will always go to the index of the list page
(scroll box)